Change directory - Change all element to components directory - Rename directory to functions
