+++ src/container/App.js
@@ -1,8 +1,10 @@
 import 'css-reset-and-normalize';
 import '../styles/global.css';
-import createElementContainer from '../elements/element_container';
+import createElementContainer from '../components/element_container';
 import Header from '../components/header';
 import Home from '../pages/home';
+import About from '../pages/about';
+import displayPage from '../functions/display';
 
 const App = () => {
   const bodyContainerElement = createElementContainer('div', [
@@ -13,7 +15,9 @@ const App = () => {
   // check with switch to change the main page
   // first check with hide and show
   // then could try deleting all inner element inerhtml=''
+  const pageToAppend = displayPage();
   main.append(Home());
+  main.append(About());
 
   bodyContainerElement.append(navBar, main);
   return bodyContainerElement;
