+++ src/functions/display.js
@@ -1,0 +1,4 @@
+const displayPage = () => {
+  return '';
+};
+export default displayPage;
