+++ src/functions/remove_elements.js
